Use 'allowHappyPath' pattern to DRY up testing of start command
I removed the gomock.InOrder so the test doesn't know the AST of the
implementation, but left in that the trust and target should
happen after Start. We lose out on testing that we're checking
the vm / virtualbox driver is compatible before calling start,
but I feel that the integration tests cover that enough.

[#134586993]